requirements.txt
alembic==1.18.4
androguard==4.1.3
apkInspector==1.3.6
# Editable install with no version control (apkshield==2.1.0)
-e /home/xssrat/projects/apkshield_v2.1
asn1crypto==1.5.1
asttokens==3.0.1
banal==1.0.6
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dataset==1.6.2
decorator==5.2.1
executing==2.2.1
fonttools==4.62.0
frida==17.8.0
greenlet==3.3.2
iniconfig==2.3.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kiwisolver==1.5.0
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mutf8==1.0.6
networkx==3.6.1
numpy==2.4.3
packaging==26.0
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
pydot==4.0.1
Pygments==2.19.2
pyOpenSSL==25.3.0
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
reportlab==4.4.10
setuptools==82.0.1
six==1.17.0
SQLAlchemy==1.4.54
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.6.0
